--- src/components/form-elements/Radio.chromatic.stories.jsx
@@ -1,6 +1,0 @@
-import * as Radio from './Radio.stories.jsx';
-import {mergeStories} from '../../chromatic/utils';
-
-export const Default = mergeStories(Radio);
-
-export default Radio.default;
