+++ src/components/form-elements/radio/_radio-group.scss
@@ -5,7 +5,7 @@
     &-column {
       flex-direction: column;
 
-      > .sg-radio-new {
+      > .sg-radio {
         margin-bottom: 8px;
       }
     }
