+++ src/docs/components/DocsActiveBlockSettings.jsx
@@ -1,7 +1,7 @@
 // @flow
 
 import ActionList from 'action-list/ActionList';
-import Radio from 'form-elements/Radio';
+import Radio from 'form-elements/radio/Radio';
 import * as React from 'react';
 import SeparatorVertical from 'separators/SeparatorVertical';
 import Text, {TEXT_SIZE} from 'text/Text';
