+++ src/sass/main.scss
@@ -26,7 +26,8 @@ $sgFontsPath: 'fonts/' !default;
 @import '../components/buttons/buttons';
 @import '../components/buttons/buttons_solid_round';
 @import '../components/icon-as-button/icon-as-button';
-@import '../components/form-elements/radio';
+@import '../components/form-elements/radio/radio';
+@import '../components/form-elements/radio/radio-group';
 @import '../components/form-elements/select';
 @import '../components/form-elements/text-input';
 @import '../components/form-elements/textarea';
@@ -63,5 +64,3 @@ $sgFontsPath: 'fonts/' !default;
 @import '../components/dialog/dialog';
 @import '../components/dialog/dialog-no-scroll';
 @import '../components/transition/transition';
-@import '../components/radio/radio';
-@import '../components/radio/radio-group';
